+++ client/public/index.html
@@ -13,7 +13,7 @@
   <script src="https://kit.fontawesome.com/f53c294a22.js"></script>
   <link href="https://fonts.googleapis.com/css?family=Fira+Sans|Satisfy&display=swap" rel="stylesheet">
 
-  <title>Art Site</title>
+  <title>Teal Canvas</title>
 </head>
 
 <body>
