+++ client/src/App.js
@@ -7,7 +7,11 @@ import AppRouter from './utilities/Router'
 function App() {
   return (
     <div className="App">
+      <div className="grid-container">
+
       <AppRouter />
+      </div>
+      
     </div>
   );
 }
