+++ client/src/components/Navbar/Navbar.css
@@ -1,34 +1,29 @@
-.grid-container {
-    width: 1000 px auto;
-    margin: 0px 0px;
-    display: grid;
-    grid-template-rows: repeat(20, 1fr) 50px;
-    grid-template-columns: repeat(7, 1fr) 100px;
-    grid-gap: 20px;
-}
+
 .logo-grid {
     grid-row: 1;
-    grid-row-end: span 2;
+    grid-row-end: span 1;
     grid-column: 1;
     grid-column-end: span 1;
     height: 100px;
     margin-left: 25px;
+    margin-top: -50px;
 }
 .nav-btn{
     color: #818181;
 }
 .navbtn-grid{
-    grid-row: 3;
+    grid-row: 2;
     grid-row-end: span 1;
     margin-left: 25px;
 }
 .nav-grid {
-    grid-row: 4;
+    grid-row: 3;
     grid-row-end: span 8;
     grid-column: 1;
     grid-column-end: span 1;
     background-color: rgba(0,0,0,0.1);
     margin-left: 25px;
+    pointer-events: none;
 }
 
 .nav-grid .router-link {
@@ -36,7 +31,8 @@
     text-decoration: none;
     font-size: 25px;
     color: #818181;
-    display: block;
+    display: flex;
+    pointer-events: none;
 }
 
 .nav-grid .router-link:hover {
