+++ client/src/utilities/Router/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Navbar from '../../components/Navbar/Navbar'
 // import * as Home from '../../pages/home'
-// import * as About from '../../pages/about'
+ import About from '../../pages/about'
 // import * as Artists from '../../pages/artists'
 // import * as Artwork from '../../pages/artwork'
 // import * as Contact from '../../pages/contact'
@@ -40,19 +40,23 @@ import Navbar from '../../components/Navbar/Navbar'
 
 function AppRouter() {
     return (
+     
         <Router>
+           
             <Navbar />
 
             <Switch>
             {/* <Route path="/login" exact component={LoginPage} /> */}
-                {/* <Route path="/" exact component={HomePage} />
-                <Route path="/about" component={AboutPage} />
-                <Route path="/artists" component={ArtistsPage} />
+                {/* { <Route path="/" exact component={HomePage} /> */}
+                <Route path="/about" component={About} />
+                {/* <Route path="/artists" component={ArtistsPage} />
                 <Route path="/artwork" component={ArtworkPage} />
                 <Route path="/contact" component={ContactPage} />
-                <Route path="/shopcart" component={ShoppingCartPage} /> */}
+                <Route path="/shopcart" component={ShoppingCartPage} />} */}
             </Switch>
+         
         </Router>
+        
     );
 }
